Move fromLine() in Group.test.ts to TestHelpers.ts
